add need to detect errors
For high-availability, exceptions are a no-go (a bad log statement can't crash the entire app).

But on the same note, it needs to be easy for developers to find bad formats in development, and some applications may even want
to format user-provided strings and so need a way to check if an error occurred. formatxx currently doesn't support those needs.