Add a Visual Studio "Shared Project" to the formatxx source.
This is not a static library due to the inability to control where static libraries are built or what compilation options are used when
importing them. Shared items instead use the configuration of the including project. This will cause the source to be needlessly
multiply-compiled and linked in some project structures, but the source is small enough here to not worry me overly.