GCC doesn't like the __cdecl on main, so do some macro magic.
This isn't necessary in the default build but turns out to be useful in my custom builds using this project, where we blanket compile
everything with /Gv on Windows (setting calling conventions to __vectorcall) and include formatxx as a subdirectory. MSVC will
complain that main't isn't __cdecl instead of just treating it as such, which is annoying, but nothing I can do about.