Replace use of decltype(args) with a variant of decay_t<Args>
std::decay is giving an unexpected and incorrect result on MS' stdlib. libstdc++ seems to be
giving incorrect results on its own though, and irritatingly only on Travis and not locally with
the same compiler version (gcc 6.3) where everything is working as expected. Basically
flailing around a bit trying to get Travis to play nice since I can't repro the bug locally.