Simplify the format forwarding wrapper.
This passes all tests on MSC and seems to maybe address the problem on GCC as evidenced by this test case:
https://godbolt.org/g/aQHLxF
Note on that test that GCC generates what appears to be "wrong" code for `test` but correctly for `simplify` while
Clang generates the expected code for both functions. This definitely looks like the failure we're seeing on
Travis. I am still unclear why it works as expected locally, and also am unclear if this is a compiler bug or a bug
in formatxx, but in either case I think this is the fix. Hopefully.