DLL support.
Export the extern templates and set a __stdcall API everywhere that linkage might matter.

The CMakeLists now has both a formatxx (shared) and formatxx_static (... static). Setting the defines
FORMATXX_EXPORT or FORMATXX_STATIC will control the headers and how/if exports happen.

I'd be surprised if this passes GCC/Clang CI on the first try.