Cleanup the formatxx headers a bit with some light and tasteful application of templates or extended template parameters.
The biggest change is that we now take format strings as a generic argument and convert it using a new make_string_view function template.
This allows other types not compatible with formatxx::string_view to be used, e.g. custom string types, by adding an overload. It also removes
the need for special overloads for std::string or even std::string_view (I do wish we could just use that and get rid of formatxx::string_view already).