Rename sformat and sprintf to format_string and printf_string, respectively.
This is mostly because some headers on Microsoft's compiler are deprecating sprintf, which poisons
the _identifier_ sprintf and not just a specific function. We can't even call a function sprintf - in any
namespace - which triggering warning 4995 on some setups. Disabling the warning in our headers
does nothing because the warning will still be triggered in user code calling our function. Dumb.
Renamed sprintf to something unique and renamed sformat for consistency to avoid the problem.